Add simple healthchecking of http endpoints (#2)
The watchdog finds healthcheck endpoints by executing FindHealthcheckEndpointsQuery query. This query looks for all services running on service fabric, and checks if they have the watchdog extension.
If a service has the watchdog extension. the query will return all the service instances, and the urls to hit the service instance healthchecks for that service.

For each healthcheck url returned by the query the watchdog will hit the url and check for a 200 response. If it does get 200 response it will report healthy to service fabric for that service instance, otherwise it will report unhealthy.

This whole process happens every 10 seconds which is setup as a scheduled job with quartz.
Need to find a good way to unit test the query.